Can promote into any of the most powerful pieces
Previously, you could only promote to a queen or rook. Now, you can promote to a queen, griffin, bombard, unicorn, or dragon rider. These are all of the most powerful pieces in the game and should infer a major advantage to a player that gets a piece across, but not an unbeatable one- I left out the minotaur and camel, which have unique enough movement patterns that it would be almost impossible to keep a king away from the moves of all of those pieces.