Added methods and variables to GameCenterHelper
These methods keep track of the total number of wins and loses. It also creates an array of them and sorts to find the highest. It then is connected to two different leader oards with the wins and loses. Some of the methods are commented out because they need further work