Corrected leaderboard code
I have changed the leaderboard code in the showLeaderBoard method in the game center helper class so it can display. I also called it in the statsButtonPressed method in the opening screen class. But for some reason It is not displaying the leader boards when you press the states button so I need to fix that